requirements.txt
flask==2.3.3
flask-pymongo==2.3.0
authlib==1.2.1
yfinance==0.2.28
python-dotenv==1.0.0
requests==2.31.0